+++ src/fields/mod.rs
@@ -1,4 +1,5 @@
 pub mod any;
+pub mod bool_literal;
 pub mod bool_type;
 pub mod collection;
 pub mod date;